Update LikedItems Model

+++ likes/models.py
@@ -8,5 +8,5 @@ class LikedItems(models.Model):
     user = models.ForeignKey(User,on_delete=models.CASCADE)
 
     content_type = models.ForeignKey(ContentType,on_delete=models.CASCADE)
-    object_id = models.IntegerField()
+    object_id = models.PositiveBigIntegerField()
     content_object = GenericForeignKey() #retrive all the liked objects from a specific person
